Remove more unused modals, de-dupe labware stack list and liquid card list
