--- app/src/molecules/LabwareStackModal/index.ts
@@ -1,1 +1,0 @@
-export * from './LabwareStackModal'
