+++ app/src/molecules/LiquidDetailCard/LiquidDetailCard.tsx
