+++ app/src/molecules/LiquidDetailCard/__tests__/LiquidDetailCard.test.tsx
