+++ app/src/molecules/LiquidDetailCard/index.ts
@@ -1,1 +1,2 @@
 export * from './LiquidDetailCard'
+export * from './LiquidCardList'
