+++ app/src/organisms/ODD/ProtocolSetup/index.ts
@@ -1,7 +1,6 @@
 export * from './ProtocolSetupDeckConfiguration'
 export * from './ProtocolSetupInstruments'
 export * from './ProtocolSetupLabware'
-export * from './ProtocolSetupLiquids'
 export * from './ProtocolSetupModulesAndDeck'
 export * from './ProtocolSetupOffsets'
 export * from './ProtocolSetupParameters'
